Put a bow on it since it appears to work now

+++ scripts/yarn/export.js
@@ -3,12 +3,15 @@ import fs from 'fs'
 import path from 'path'
 import { fileURLToPath } from 'url'
 
-const __filename = fileURLToPath(import.meta.url)
-const __dirname = path.dirname(__filename)
+// Clean the export directory to get tugboat working
+if (process.env.BUILD_OPTION === 'static') {
+  const __filename = fileURLToPath(import.meta.url)
+  const __dirname = path.dirname(__filename)
 
-const exportBuildPath = path.resolve(__dirname, '../../.next/export')
-if (fs.existsSync(exportBuildPath)) {
-  fs.rmdirSync(exportBuildPath, { recursive: true, force: true })
+  const exportBuildPath = path.resolve(__dirname, '../../.next/export')
+  if (fs.existsSync(exportBuildPath)) {
+    fs.rmdirSync(exportBuildPath, { recursive: true, force: true })
+  }
 }
 
 processEnv('next build')
